back/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8.5,<3.0
factory_boy<=3.2
requests<=2.26.0
coverage<=5.5
django-extensions>=3.0.5
djangorestframework>=3.12.2
django-cors-headers>=3.8.0
